requirements.txt
pytest==6.2.5
pytest-html==3.1.1
pytest-xdist==2.4.0
pytest-ordering==0.6
pytest-rerunfailures==10.2
allure-pytest==2.9.45
loguru==0.5.3
configparser==5.2.0
PyHamcrest==2.0.3
jsonpath==0.82
PyMySQL==1.0.2
prettytable==3.2.0
PyYAML==6.0